+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.6.1](https://github.com/crowdsecurity/php-capi-client/releases/tag/v0.6.1) - 2022-12-16
+[_Compare with previous release_](https://github.com/crowdsecurity/php-capi-client/compare/v0.6.0...v0.6.1)
+
+### Fixed
+
+- Fix `Curl` unlimited timeout when negative value is configured in `api_timeout`
+
+
+---
+
 ## [0.6.0](https://github.com/crowdsecurity/php-capi-client/releases/tag/v0.6.0) - 2022-12-16
 [_Compare with previous release_](https://github.com/crowdsecurity/php-capi-client/compare/v0.5.0...v0.6.0)